+++ tests/input_files/plot_options/treegraph/01.js
@@ -1,0 +1,22 @@
+{
+  type: 'treegraph',
+  animationLimit: 10,
+  boostBlending: 'some-value-goes-here',
+  boostThreshold: 5000,
+  colorIndex: 2,
+  crisp: true,
+  cropThreshold: 500,
+  findNearestPointBy: 'xy',
+  getExtremesFromAll: true,
+  relativeXValue: true,
+  softThreshold: true,
+  step: 'left',
+
+  pointInterval: 5,
+  pointIntervalUnit: 'day',
+  pointStart: 1,
+  stacking: 'normal',
+
+  allowTraversingTree: true,
+  colorByPoint: true
+}